fix search following pull request comments
