+++ src/Search/Input/index.js
@@ -1,0 +1,4 @@
+// @flow
+import Input from "./Input";
+
+export default Input;
